Fix off-by-one errors in channel.readline.
This fixes the following bugs in channel.readline:
1. The default value expression for the ``amount`` argument resulted in 1 less than the true
amount. This meant that arrays of length 1 could not be used. This does not affect users who
specify their own ``amount``. I have added 1 to the calculation of amount.
2. The documentation used to say that the newline was not included in the array. I have
changed the documentation to match the existing behaviour (this also matches the behaviour of
the string version of readline) since at least some code relied on it. (for example test/io/bharshbarg/readlineArray.chpl)
3. The return value of qio_channel_read_byte was always written into the array (and ``numRead`` incremented), even when what was returned was an error code. Now it does not write
to the array or increment ``numRead`` unless there was no error.
4. If EOF is encountered before ``\n``, but bytes had been read an error would be returned. I
have changed it so that ``error`` would only be set to EEOF when no bytes were read.
5. The version of readline with no error= is supposed to call _ch_ioerror for an error message
on all errors except EEOF which should just return false. I have removed the test for !got
since that would always be true when there is an error, causing the error message prining
branch to be unreachable.